Inherit `default-features = false` from packages
When packages are added to the workspace they seem to infer default
features even if all workspace members disable default features: hence
this flag must be propagated to the workspace dependencies list.

It should not be necessary to inherit feature flags from individual
crates though, as these are additive across the workspace (but it might
be cleaner to strip all common feature flags?).